Solving Sudoku problem using Constraint Satisfaction Search
You need to solve the 9x9 Sudoku as Constraint Satisfaction Problem as discussed in class on  Sept. 16, 2021.
You need to define the structure and constraints for this problem. 
Also problem should be defined properly on paper before coding.
A running program also needs to be shown after submission.
Try to code as an AI technique, if possible, meaning that you can generalize this for 16x16 or some other size like 25x25,  36x36
You can do it in a group of max 2 persons.